requirements.txt
# Core framework
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# CLI interface
rich>=13.7.0
prompt-toolkit>=3.0.43

# Tools & utilities
gitpython>=3.1.40
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0

# Code analysis
radon>=6.0.1
